Fix signed integer overflow in RV32I
The implementation of add, sub, and addi instructions incorrectly used
int32_t for arithmetic operations, leading to signed integer overflow.
Address the issue by maintaining the uint32_t type for arithmetic
operations, ensuring compliance with the laws of arithmetic modulo 2^n.
This approach prevents undefined behavior resulting from signed integer
overflow.

+++ src/rv32_constopt.c
@@ -233,8 +233,7 @@ CONSTOPT(srai, {
 CONSTOPT(add, {
     if (info->is_constant[ir->rs1] && info->is_constant[ir->rs2]) {
         info->is_constant[ir->rd] = true;
-        ir->imm = (int32_t) info->const_val[ir->rs1] +
-                  (int32_t) info->const_val[ir->rs2];
+        ir->imm = info->const_val[ir->rs1] + info->const_val[ir->rs2];
         info->const_val[ir->rd] = ir->imm;
         ir->opcode = rv_insn_lui;
         ir->impl = dispatch_table[ir->opcode];
@@ -246,8 +245,7 @@ CONSTOPT(add, {
 CONSTOPT(sub, {
     if (info->is_constant[ir->rs1] && info->is_constant[ir->rs2]) {
         info->is_constant[ir->rd] = true;
-        ir->imm = (int32_t) info->const_val[ir->rs1] -
-                  (int32_t) info->const_val[ir->rs2];
+        ir->imm = info->const_val[ir->rs1] - info->const_val[ir->rs2];
         info->const_val[ir->rd] = ir->imm;
         ir->opcode = rv_insn_lui;
         ir->impl = dispatch_table[ir->opcode];
